added details
